The table name of students used trial lesson has been corrected and the final refactoring has been performed.
