+++ src/main/java/com/ieltswise/IELTSWiseApplication.java
@@ -1,4 +1,4 @@
-package com.ietswise;
+package com.ieltswise;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
