+++ src/main/java/com/ieltswise/entity/Event.java
@@ -1,4 +1,4 @@
-package com.ietswise.entity;
+package com.ieltswise.entity;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
