+++ src/main/java/com/ieltswise/entity/FreeAndBusyHoursOfTheDay.java
@@ -1,4 +1,4 @@
-package com.ietswise.entity;
+package com.ieltswise.entity;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
