+++ src/main/java/com/ieltswise/entity/StudentUsedTrial.java
@@ -1,4 +1,4 @@
-package com.ietswise.entity;
+package com.ieltswise.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
