+++ src/main/java/com/ieltswise/repository/StudentRepository.java
@@ -1,6 +1,6 @@
-package com.ietswise.repository;
+package com.ieltswise.repository;
 
-import com.ietswise.entity.Student;
+import com.ieltswise.entity.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StudentRepository extends JpaRepository<Student, Long> {
